+++ src/wallets/web3-provider/providers/ws-provider.js
@@ -13,14 +13,11 @@ import {
 class WSProvider {
   constructor(host, options, store, eventHub) {
     this.wsProvider = new Web3WSProvider(host, options);
-    const _this = this.wsProvider;
-    const requestManager = new Web3RequestManager(
-      new Web3WSProvider(host, options)
-    );
     delete this.wsProvider['send'];
     this.wsProvider.send = (payload, callback) => {
+      const _this = this.wsProvider;
       if (_this.connection.readyState === _this.connection.CONNECTING) {
-        setTimeout(() => {
+        setTimeout(function() {
           this.wsProvider.send(payload, callback);
         }, 10);
         return;
@@ -35,7 +32,9 @@ class WSProvider {
       const req = {
         payload,
         store,
-        requestManager,
+        requestManager: new Web3RequestManager(
+          new Web3WSProvider(host, options)
+        ),
         eventHub
       };
       const middleware = new MiddleWare();